+++ src/config.rs
@@ -250,6 +250,10 @@ pub struct HookBuilder {
     display_env_section: bool,
     panic_section: Option<Box<dyn Display + Send + Sync + 'static>>,
     panic_message: Box<dyn PanicMessage>,
+    #[cfg(feature = "issue-url")]
+    issue_url: Option<String>,
+    #[cfg(feature = "issue-url")]
+    issue_metadata: Vec<(String, Box<dyn Display + Send + Sync + 'static>)>,
 }
 
 impl HookBuilder {
@@ -284,6 +288,10 @@ impl HookBuilder {
             display_env_section: true,
             panic_section: None,
             panic_message: Box::new(DefaultPanicMessage),
+            #[cfg(feature = "issue-url")]
+            issue_url: None,
+            #[cfg(feature = "issue-url")]
+            issue_metadata: vec![],
         }
     }
 
@@ -360,6 +368,57 @@ impl HookBuilder {
         self
     }
 
+    /// Set an upstream github repo and enable issue reporting url generation
+    ///
+    /// # Details
+    ///
+    /// Once enabled, color-eyre will generate urls that will create customized
+    /// issues pre-populated with information about the associated error report.
+    ///
+    /// Additional information can be added to the metadata table in the
+    /// generated urls by calling `add_issue_metadata` when configuring the
+    /// HookBuilder.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// color_eyre::config::HookBuilder::default()
+    ///     .issue_url("https://github.com/yaahc/jane-eyre/issues/new")
+    ///     .install()
+    ///     .unwrap();
+    /// ```
+    #[cfg(feature = "issue-url")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "issue-url")))]
+    pub fn issue_url<S: ToString>(mut self, url: S) -> Self {
+        self.issue_url = Some(url.to_string());
+        self
+    }
+
+    /// Add a new entry to the metadata table in generated github issue urls
+    ///
+    /// **Note**: this metadata will be ignored if no `issue_url` is set.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// color_eyre::config::HookBuilder::default()
+    ///     .issue_url("https://github.com/yaahc/jane-eyre/issues/new")
+    ///     .add_issue_metadata("version", "0.1.0")
+    ///     .install()
+    ///     .unwrap();
+    /// ```
+    #[cfg(feature = "issue-url")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "issue-url")))]
+    pub fn add_issue_metadata<K, V>(mut self, key: K, value: V) -> Self
+    where
+        K: Display,
+        V: Display + Send + Sync + 'static,
+    {
+        let pair = (key.to_string(), Box::new(value) as _);
+        self.issue_metadata.push(pair);
+        self
+    }
+
     /// Configures the default capture mode for `SpanTraces` in error reports and panics
     pub fn capture_span_trace_by_default(mut self, cond: bool) -> Self {
         self.capture_span_trace_by_default = cond;
@@ -423,6 +482,8 @@ impl HookBuilder {
     }
 
     pub(crate) fn into_hooks(self) -> (PanicHook, EyreHook) {
+        #[cfg(feature = "issue-url")]
+        let metadata = Arc::new(self.issue_metadata);
         let panic_hook = PanicHook {
             filters: self.filters.into_iter().map(Into::into).collect(),
             section: self.panic_section,
@@ -430,12 +491,20 @@ impl HookBuilder {
             capture_span_trace_by_default: self.capture_span_trace_by_default,
             display_env_section: self.display_env_section,
             panic_message: self.panic_message,
+            #[cfg(feature = "issue-url")]
+            issue_url: self.issue_url.clone(),
+            #[cfg(feature = "issue-url")]
+            issue_metadata: metadata.clone(),
         };
 
         let eyre_hook = EyreHook {
             #[cfg(feature = "capture-spantrace")]
             capture_span_trace_by_default: self.capture_span_trace_by_default,
             display_env_section: self.display_env_section,
+            #[cfg(feature = "issue-url")]
+            issue_url: self.issue_url,
+            #[cfg(feature = "issue-url")]
+            issue_metadata: metadata,
         };
 
         (panic_hook, eyre_hook)
@@ -533,10 +602,17 @@ fn print_panic_info(printer: &PanicPrinter<'_>, out: &mut fmt::Formatter<'_>) ->
     hook.panic_message.display(printer.0, out)?;
 
     let v = panic_verbosity();
+    let capture_bt = v != Verbosity::Minimal;
 
     #[cfg(feature = "capture-spantrace")]
     let span_trace = if hook.spantrace_capture_enabled() {
         Some(tracing_error::SpanTrace::capture())
+    } else {
+        None
+    };
+
+    let bt = if capture_bt {
+        Some(backtrace::Backtrace::new())
     } else {
         None
     };
@@ -558,10 +634,7 @@ fn print_panic_info(printer: &PanicPrinter<'_>, out: &mut fmt::Formatter<'_>) ->
         }
     }
 
-    let capture_bt = v != Verbosity::Minimal;
-
-    if capture_bt {
-        let bt = backtrace::Backtrace::new();
+    if let Some(bt) = bt.as_ref() {
         let fmted_bt = hook.format_backtrace(&bt);
         write!(
             indented(&mut separated.ready()).with_format(Format::Uniform { indentation: "  " }),
@@ -580,6 +653,27 @@ fn print_panic_info(printer: &PanicPrinter<'_>, out: &mut fmt::Formatter<'_>) ->
         write!(&mut separated.ready(), "{}", env_section)?;
     }
 
+    #[cfg(feature = "issue-url")]
+    if let Some(url) = &hook.issue_url {
+        let payload = printer
+            .0
+            .payload()
+            .downcast_ref::<String>()
+            .map(String::as_str)
+            .or_else(|| printer.0.payload().downcast_ref::<&str>().cloned())
+            .unwrap_or("<non string panic payload>");
+
+        let issue_section = crate::section::github::IssueSection::new(url, payload)
+            .with_backtrace(bt.as_ref())
+            .with_location(printer.0.location())
+            .with_metadata(&**hook.issue_metadata);
+
+        #[cfg(feature = "capture-spantrace")]
+        let issue_section = issue_section.with_span_trace(span_trace.as_ref());
+
+        write!(&mut separated.ready(), "{}", issue_section)?;
+    }
+
     Ok(())
 }
 
@@ -590,6 +684,10 @@ pub(crate) struct PanicHook {
     #[cfg(feature = "capture-spantrace")]
     capture_span_trace_by_default: bool,
     display_env_section: bool,
+    #[cfg(feature = "issue-url")]
+    issue_url: Option<String>,
+    #[cfg(feature = "issue-url")]
+    issue_metadata: Arc<Vec<(String, Box<dyn Display + Send + Sync + 'static>)>>,
 }
 
 impl PanicHook {
@@ -615,6 +713,10 @@ pub(crate) struct EyreHook {
     #[cfg(feature = "capture-spantrace")]
     capture_span_trace_by_default: bool,
     display_env_section: bool,
+    #[cfg(feature = "issue-url")]
+    issue_url: Option<String>,
+    #[cfg(feature = "issue-url")]
+    issue_metadata: Arc<Vec<(String, Box<dyn Display + Send + Sync + 'static>)>>,
 }
 
 impl EyreHook {
@@ -641,6 +743,10 @@ impl EyreHook {
             span_trace,
             sections: Vec::new(),
             display_env_section: self.display_env_section,
+            #[cfg(feature = "issue-url")]
+            issue_url: self.issue_url.clone(),
+            #[cfg(feature = "issue-url")]
+            issue_metadata: self.issue_metadata.clone(),
         }
     }