+++ src/section/mod.rs
@@ -2,6 +2,8 @@
 use crate::writers::WriterExt;
 use std::fmt::{self, Display};
 
+#[cfg(feature = "issue-url")]
+pub(crate) mod github;
 pub(crate) mod help;
 
 /// An indented section with a header for an error report